choose layout buttons svg
